+++ laoschoolws/src/main/java/com/itpro/restws/dao/MUser2ClassDaoImpl.java
@@ -16,19 +16,19 @@ public class MUser2ClassDaoImpl extends AbstractDao<Integer, MUser2Class> implem
 
 	
 	final String ModelName = "MUser2Class";
-	public MUser2Class findById(int id) {
-		return getByKey(id);
+	public MUser2Class findById(Integer id) {
+		return getByKey(id.intValue());
 	}
 
 	@Override
-	public int countBySchool(int school_id) {
+	public int countBySchool(Integer school_id) {
 		// Get row count
-		int count = ((Long)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
+		int count = ((Integer)getSession().createQuery("select count(*) from " + ModelName+  " WHERE school_id = '" + school_id + "'").uniqueResult()).intValue();
 		return count;
 	}
 
 	@Override
-	public List<MUser2Class> findBySchool(int school_id, int from_row, int max_result) {
+	public List<MUser2Class> findBySchool(Integer school_id, int from_row, int max_result) {
 		Criteria crit_list = createEntityCriteria();
 		crit_list.add(Restrictions.eq("school_id", school_id));
 		crit_list.setMaxResults(max_result);
